Only send certain env headers to papermill.
JSON serialization was blowing up previously, because we were calling
.to_json on the entire env hash, which includes IO objects (such as
rack.errors), which are not opened for reading when serialization occurs.
The solution is to not serialize the entire env hash, instead grabbing
only the fields that we actually care about.

This commit also includes fixing certain env headers key names. Rack
prefixes some headers with rack.*, which is not a valid key name in
mongodb, thus, we need to change the key names.